NRU, LRU works

+++ virtual-memory/include/list.h
@@ -10,7 +10,7 @@ typedef struct node {
 } Node;
  
 typedef struct list {
-  size_t sz;
+  int sz;
   Node* head;
   Node* last;
 } List;
@@ -20,6 +20,5 @@ void list_push(List* list, void* data);
 void* list_pop(List* list);
 void list_free(List* list);
 void list_print(List* list);
-
 
 #endif
